Base xcode project
This has been set up with a generic organization name and no team identity.
Since we don't have a legal entity company, we can't register as an Apple Developer organization.
Later we can decide if we want one person with an Apple Developer Program membership to push this to the app store.