Show results in a table for development/debugging.
The whole login storyboard is throwaway, but I am creating a table to
display search results and debug sorting/filtering of our places array.
We want to optimize our sort so that we can choose the best 5-10 places
to display on our map. Once sort is optimized, we can unify on that
sort function and hide access to this storyboard.